Test suite is built with cmake on both x86_64 and RISC-V
