+++ test/intercept_sys_mkdirat.c
@@ -30,55 +30,37 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#define _GNU_SOURCE
-#include <sched.h>
-#include <stdio.h>
+#include "libsyscall_intercept_hook_point.h"
+
+#include <syscall.h>
 #include <unistd.h>
-#include <sys/types.h>
-#include <sys/wait.h>
-#include <assert.h>
-#include <fcntl.h>
+#include <stdio.h>
 #include <string.h>
-#include <stdlib.h>
+#include <stdint.h>
 
-int child_func(void *arg) {
-    int fd = openat(AT_FDCWD, "testfile.txt", O_RDONLY);
-    char buf[128];
-    int n = read(fd, buf, sizeof(buf));
-    buf[n] = '\0';
-    n = atoi(buf);
-    assert(n == getpid());
-    return 0;
+
+static int hook(long syscall_number,
+                long arg0, long arg1,
+                long arg2, long arg3,
+                long arg4, long arg5,
+                long *result)
+{
+    (void) result;
+
+    if (syscall_number == SYS_mkdirat) {
+        char buf[128] = "wrongdir/";
+        char *tmp = buf;
+        if (strcmp((char *)arg1, tmp) == 0) {
+            const char testdir[] = "testdir/";
+            syscall_no_intercept(syscall_number, arg0, (uintptr_t)testdir, arg2,
+                                 arg3, arg4, arg5);
+            return 0;
+        }
+    }
+    return 1;
 }
 
-int main() {
-    int fd = openat(AT_FDCWD, "testfile.txt", O_CREAT | O_TRUNC | O_RDWR, 0666);
-    int fd2 = openat(AT_FDCWD, "testfile2.txt", O_CREAT | O_TRUNC | O_RDWR, 0666);
-
-    char child_stack[8192];
-
-    pid_t pid = clone(child_func, child_stack + sizeof(child_stack),
-                      SIGCHLD, NULL);
-
-    if (pid == -1) {
-        perror("Clone failed");
-        return 1;
-    }
-
-    int status;
-    wait(&status);
-    if (WIFSIGNALED(status) && WTERMSIG(status) == SIGABRT) {
-        fprintf(stderr, "Child assertion failed\n");
-        return 1;
-    }
-
-    char buf[128];
-    int n = read(fd2, buf, sizeof(buf));
-    buf[n] = '\0';
-    n = atoi(buf);
-    assert(n == getpid());
-    write(1, "CLONE TEST - OK\n", 16);
-    return 0;
+static __attribute__((constructor)) void init(void)
+{
+    intercept_hook_point = hook;
 }
-
-
