+++ test/write_test.c
